+++ docs/CHANGELOG.md
@@ -4,7 +4,16 @@ All notable changes to `newrelic` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## 2018-01-25
+
+### Added
+- `NewRelic` and `NewRelicBlackhole` classes, which implement `NewRelicInterface`
+- `NewRelicFactory` which is static factory to get an instance of `NewRelicInterface`
+- Test to check proper initialization using `NewRelicFactory`
+
+***
+
 ## 2017-11-19
 
 ### Added
-- Nothing
+- `NewRelicInterface`